feat: break out client state machine
