+++ packages/studio-base/src/panels/Plot/useDatasets.ts
@@ -12,42 +12,34 @@ import { Immutable } from "@foxglove/studio";
 import { useMessageReducer as useCurrent, useDataSourceInfo } from "@foxglove/studio-base/PanelAPI";
 import { useBlocksSubscriptions as useBlocks } from "@foxglove/studio-base/PanelAPI/useBlocksSubscriptions";
 import {
-  RosPath,
-  MessagePathPart,
-} from "@foxglove/studio-base/components/MessagePathSyntax/constants";
-import parseRosPath from "@foxglove/studio-base/components/MessagePathSyntax/parseRosPath";
-import {
   useMessagePipeline,
   MessagePipelineContext,
 } from "@foxglove/studio-base/components/MessagePipeline";
-import { mergeSubscriptions } from "@foxglove/studio-base/components/MessagePipeline/subscriptions";
 import { TypedDataProvider } from "@foxglove/studio-base/components/TimeBasedChart/types";
 import useGlobalVariables from "@foxglove/studio-base/hooks/useGlobalVariables";
 import { SubscribePayload, MessageEvent } from "@foxglove/studio-base/players/types";
 
 import {
-  initBlockState,
-  refreshBlockTopics,
-  processBlocks,
-  prepareBlockUpdate,
-  ClientUpdate,
-  BlockState,
-} from "./blocks";
+  DatasetsState,
+  initDatasets,
+  registerClient,
+  unregisterClient,
+  updateParams,
+  getSubscriptions,
+  updateBlocks,
+  resetClientBlocks,
+  splitSubscriptions,
+} from "./clients";
 import { PlotParams } from "./internalTypes";
-import { getPaths } from "./params";
 import { PlotData } from "./plotData";
 
 type Service = Comlink.Remote<(typeof import("./useDatasets.worker"))["service"]>;
-type Client = {
-  params: PlotParams | undefined;
-  setter: (topics: SubscribePayload[]) => void;
-  blocks: BlockState;
-};
 
 let worker: Worker | undefined;
 let service: Service | undefined;
 let numClients: number = 0;
-let clients: Record<string, Client> = {};
+let datasetsState: DatasetsState = initDatasets();
+let callbacks: Record<string, (topics: SubscribePayload[]) => void> = {};
 
 const pending: ((service: Service) => void)[] = [];
 async function waitService(): Promise<Service> {
@@ -61,154 +53,42 @@ async function waitService(): Promise<Service> {
 
 const getIsLive = (ctx: MessagePipelineContext) => ctx.seekPlayback == undefined;
 
-/**
- * Get the SubscribePayload for a single path by subscribing to all fields
- * referenced in leading MessagePathFilters and the first field of the
- * message.
- */
-export function pathToPayload(path: RosPath): SubscribePayload | undefined {
-  const { messagePath: parts, topicName: topic } = path;
-
-  // We want to take _all_ of the filters that start the path, since these can
-  // be chained
-  const filters = R.takeWhile((part: MessagePathPart) => part.type === "filter", parts);
-  const firstField = parts.find((part: MessagePathPart) => part.type === "name");
-  if (firstField == undefined || firstField.type !== "name") {
-    return undefined;
-  }
-
-  return {
-    topic,
-    fields: R.pipe(
-      R.chain((part: MessagePathPart): string[] => {
-        if (part.type !== "filter") {
-          return [];
-        }
-        const { path: filterPath } = part;
-        const field = filterPath[0];
-        if (field == undefined) {
-          return [];
-        }
-
-        return [field];
-      }),
-      // Always subscribe to the header field
-      (filterFields) => [...filterFields, firstField.name, "header"],
-      R.uniq,
-    )(filters),
-  };
-}
-
-function getPayloadsFromPaths(paths: readonly string[]): SubscribePayload[] {
-  return R.pipe(
-    R.chain((path: string): SubscribePayload[] => {
-      const parsed = parseRosPath(path);
-      if (parsed == undefined) {
-        return [];
-      }
-
-      const payload = pathToPayload(parsed);
-      if (payload == undefined) {
-        return [];
-      }
-
-      return [payload];
-    }),
-    // Then simplify
-    (v: SubscribePayload[]) => mergeSubscriptions(v) as SubscribePayload[],
-  )(paths);
-}
-
-function getPayloadsFromClient(client: Client): SubscribePayload[] {
-  const { params } = client;
-  if (params == undefined) {
-    return [];
-  }
-
-  const { xAxisPath, paths: yAxisPaths } = params;
-
-  return R.pipe(
-    getPayloadsFromPaths,
-    R.chain((v): SubscribePayload[] => {
-      const partial: SubscribePayload = {
-        ...v,
-        preloadType: "partial",
-      };
-
-      // Subscribe to both "partial" and "full" when using "full" In
-      // theory, "full" should imply "partial" but not doing this breaks
-      // MockMessagePipelineProvider
-      return [partial, { ...partial, preloadType: "full" }];
-    }),
-  )(getPaths(yAxisPaths, xAxisPath));
-}
-
 // Calculate the list of unique topics that _all_ of the plots need and
 // nominate one panel to subscribe to the topics on behalf of the rest.
 function chooseClient() {
-  if (R.isEmpty(clients)) {
+  if (R.isEmpty(datasetsState.clients)) {
     return;
   }
 
-  const clientList = Object.values(clients);
-  const subscriptions = R.pipe(
-    R.chain(getPayloadsFromClient),
-    (v) => mergeSubscriptions(v) as SubscribePayload[],
-  )(clientList);
-  clientList[0]?.setter(subscriptions);
+  const clientList = Object.values(callbacks);
+  clientList[0]?.(getSubscriptions(datasetsState));
 }
 
 function clearClient(id: string) {
-  clients = R.mapObjIndexed((client, clientId) => {
-    if (clientId !== id) {
-      return client;
-    }
-
-    return {
-      ...client,
-      blocks: initBlockState(),
-    };
-  }, clients);
+  const [newState, update] = resetClientBlocks(id, datasetsState);
+  datasetsState = newState;
+  void service?.addBlockData(update);
 }
 
-// Subscribe to "current" messages (those near the seek head) and forward new
-// messages to the worker as they arrive.
+// Subscribe to all of the data all plots require and forward it to the worker.
 function useData(id: string, params: PlotParams) {
   const [subscriptions, setSubscribed] = React.useState<SubscribePayload[]>([]);
   // Register client when the panel mounts and unregister when it unmounts
   useEffect(() => {
-    clients = {
-      ...clients,
-      [id]: {
-        params: undefined,
-        blocks: initBlockState(),
-        setter: setSubscribed,
-      },
-    };
+    datasetsState = registerClient(id, datasetsState);
+    callbacks[id] = setSubscribed;
     chooseClient();
     return () => {
-      const { [id]: _client, ...rest } = clients;
-      clients = rest;
+      datasetsState = unregisterClient(id, datasetsState);
+      const { [id]: _client, ...rest } = callbacks;
+      callbacks = rest;
       chooseClient();
     };
   }, [id]);
 
   // Update registration when params change
   useEffect(() => {
-    const { [id]: client } = clients;
-    if (client == undefined) {
-      return;
-    }
-
-    const { blocks } = client;
-    clients = {
-      ...clients,
-      [id]: {
-        ...client,
-        blocks: refreshBlockTopics(getPayloadsFromClient(client), blocks),
-        params,
-      },
-    };
+    datasetsState = updateParams(id, params, datasetsState);
     chooseClient();
   }, [id, params]);
 
@@ -221,7 +101,7 @@ function useData(id: string, params: PlotParams) {
   }, [isLive]);
 
   const [blockSubscriptions, currentSubscriptions] = React.useMemo(
-    () => R.partition((v) => v.preloadType === "full", subscriptions),
+    () => splitSubscriptions(subscriptions),
     [subscriptions],
   );
 
@@ -248,33 +128,9 @@ function useData(id: string, params: PlotParams) {
       return;
     }
 
-    const clientsAndUpdates = R.toPairs(clients).map(
-      ([clientId, client]): [id: string, client: Client, updates: ClientUpdate[]] => {
-        const { state: newBlocks, updates } = processBlocks(
-          blocks,
-          blockSubscriptions,
-          client.blocks,
-        );
-        const newClient = {
-          ...client,
-          blocks: newBlocks,
-        };
-        return [
-          clientId,
-          newClient,
-          updates.map((update): ClientUpdate => ({ id: clientId, update })),
-        ];
-      },
-    );
-
-    clients = R.fromPairs(clientsAndUpdates.map(([clientId, client]) => [clientId, client]));
-
-    void service?.addBlockData(
-      prepareBlockUpdate(
-        clientsAndUpdates.flatMap(([, , updates]): ClientUpdate[] => updates),
-        blocks,
-      ),
-    );
+    const [newState, blockUpdate] = updateBlocks(blocks, datasetsState);
+    datasetsState = newState;
+    void service?.addBlockData(blockUpdate);
   }, [blockSubscriptions, blocks]);
 }
